requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pyjwt==2.8.0
pytz==2022.6
sqlparse==0.5.0
tzdata==2024.1
django-oauth-toolkit==2.3.0
jwcrypto==1.5.6
google-auth-oauthlib==1.2.0